Add files via upload
In this project we classify the sentences on the basis of the their toxicity which were divided into 6 different groups. First we train out model on the data set which is taken from the Kaggle which is a large data of wikipedia comments, on the basis of this data we train our two model here logistic regression and the naive bayes after the preprocessing the data. Both model gives good accuracy but when we talk about the camparison then logistic regression performance here is much better then the naive Bayes. After all we also draw a ROC curve.